Add menu_screen, update category
