+++ lib/src/widgets/category.dart
@@ -27,7 +27,7 @@ class CategorySection extends StatelessWidget {
             crossAxisCount: 2,
             crossAxisSpacing: 8,
             mainAxisSpacing: 8,
-            childAspectRatio: 0.55,
+            childAspectRatio: 0.6,
           ),
           itemCount: coffeeItems.length,
           itemBuilder: (context, index) => CoffeeItem(
